version-python/biometrico/requirements.txt
altair==5.5.0
altgraph==0.17.5
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
fonttools==4.59.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mysql-connector-python==9.5.0
narwhals==2.8.0
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pefile==2024.8.26
pillow==11.3.0
ply==3.11
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
pyzk @ git+https://github.com/fananimi/pyzk.git@e2c391fc0a002a96386d95b23cfd8615bf25e714
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.1
selenium==4.35.0
setuptools==75.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
thriftpy2==0.5.3
toml==0.10.2
tornado==6.5.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
